backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
faster-whisper
torch
torchaudio
google-generativeai
requests
coqui-tts
python-multipart
